dev-php/composer: Add new test with own autoloader
Before the test does not work if the package was not installed,
becasuse it *needs* the autoloader, it was bad.

Now I copy the autoloader, from the portage tree FILESDIR,
then change the path according to the current WORKDIR,
with that process you can test without having it install,
or against different version.